Changed multiple packages to fit the new Data Base configuration Added TODO to AdminServiceImp, UserServiceImp Refactorized UserServiceImp to use new features Renamed some bad variable names
Removed unuseful comments and removed unused import annotations
